Added a feature to launch actual games. (in systems/[systemname]/config.ini the emu_path must be set. the concat function has been moved from filesystem::directory to filesystem::path filesystem::path now has escape(). it currently escapes \, ', (, ), ' ' game has some properties added for reading its filename. also an empty function (to be filled in) for friendly name creating. (so removing (USA)[!]_.* from filenames.
TODO:
when a game is launched from arcade::engine the program waits for the emulator to stop.
this gives a wait signal to the operating system.
I think it's better to fork or create some kind of thread (or just ignore waiting).
and check for its process to end / regain focus. using that status to set Engine::Status::Active/Paused